backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
psycopg
psycopg2-binary
python-dotenv
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
msgpack==1.1.0
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
